requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8

# Optional: For PostgreSQL/Supabase (uncomment if needed)
# psycopg[binary]==3.1.18
